Moved Polygon to its own source file
