+++ pyatlas/doc/atlas.html
@@ -21,15 +21,16 @@
 <td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="area.html">area</a><br>
 <a href="atlas_metadata.html">atlas_metadata</a><br>
 <a href="autogen.html">autogen</a><br>
-</td><td width="25%" valign=top><a href="edge.html">edge</a><br>
-<a href="integer_dictionary.html">integer_dictionary</a><br>
+<a href="edge.html">edge</a><br>
+</td><td width="25%" valign=top><a href="integer_dictionary.html">integer_dictionary</a><br>
 <a href="line.html">line</a><br>
-</td><td width="25%" valign=top><a href="node.html">node</a><br>
+<a href="node.html">node</a><br>
 <a href="packed_tag_store.html">packed_tag_store</a><br>
-<a href="point.html">point</a><br>
-</td><td width="25%" valign=top><a href="polyline.html">polyline</a><br>
+</td><td width="25%" valign=top><a href="point.html">point</a><br>
+<a href="polygon.html">polygon</a><br>
+<a href="polyline.html">polyline</a><br>
 <a href="relation.html">relation</a><br>
-<a href="zipfile.html">zipfile</a><br>
+</td><td width="25%" valign=top><a href="zipfile.html">zipfile</a><br>
 </td></tr></table></td></tr></table><p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#ee77aa">
