+++ pyatlas/doc/polygon.html
@@ -1,0 +1,100 @@
+
+<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
+<html><head><title>Python: module polygon</title>
+<meta charset="utf-8">
+</head><body bgcolor="#f0f0f8">
+
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="heading">
+<tr bgcolor="#7799ee">
+<td valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>polygon</strong></big></big></font></td
+><td align=right valign=bottom
+><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/lucascram/Projects/atlas/pyatlas/pyatlas/polygon.py">/Users/lucascram/Projects/atlas/pyatlas/pyatlas/polygon.py</a></font></td></tr></table>
+    <p></p>
+<p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#aa55cc">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="polyline.html">polyline</a><br>
+</td><td width="25%" valign=top></td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#ee77aa">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Classes</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#ee77aa"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><dl>
+<dt><font face="helvetica, arial"><a href="polyline.html#PolyLine">polyline.PolyLine</a>(boundable.Boundable)
+</font></dt><dd>
+<dl>
+<dt><font face="helvetica, arial"><a href="polygon.html#Polygon">Polygon</a>
+</font></dt></dl>
+</dd>
+</dl>
+ <p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#ffc8d8">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#000000" face="helvetica, arial"><a name="Polygon">class <strong>Polygon</strong></a>(<a href="polyline.html#PolyLine">polyline.PolyLine</a>)</font></td></tr>
+    
+<tr bgcolor="#ffc8d8"><td rowspan=2><tt>&nbsp;&nbsp;&nbsp;</tt></td>
+<td colspan=2><tt>A&nbsp;special&nbsp;case&nbsp;of&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>&nbsp;that&nbsp;has&nbsp;an&nbsp;extra&nbsp;segment&nbsp;between&nbsp;the&nbsp;last&nbsp;and<br>
+first&nbsp;point&nbsp;-&nbsp;effectively&nbsp;a&nbsp;closed&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>.&nbsp;The&nbsp;<a href="#Polygon">Polygon</a>&nbsp;does&nbsp;not&nbsp;actually<br>
+re-store&nbsp;the&nbsp;last&nbsp;(first)&nbsp;Location.&nbsp;Instead,&nbsp;the&nbsp;API&nbsp;simulates&nbsp;its&nbsp;presence.<br>&nbsp;</tt></td></tr>
+<tr><td>&nbsp;</td>
+<td width="100%"><dl><dt>Method resolution order:</dt>
+<dd><a href="polygon.html#Polygon">Polygon</a></dd>
+<dd><a href="polyline.html#PolyLine">polyline.PolyLine</a></dd>
+<dd>boundable.Boundable</dd>
+<dd><a href="__builtin__.html#object">__builtin__.object</a></dd>
+</dl>
+<hr>
+Methods defined here:<br>
+<dl><dt><a name="Polygon-__init__"><strong>__init__</strong></a>(self, location_list, deep<font color="#909090">=False</font>)</dt><dd><tt>Create&nbsp;a&nbsp;new&nbsp;<a href="#Polygon">Polygon</a>&nbsp;given&nbsp;a&nbsp;Location&nbsp;list.&nbsp;By&nbsp;default,&nbsp;it&nbsp;will&nbsp;perform<br>
+a&nbsp;shallow&nbsp;copy&nbsp;of&nbsp;the&nbsp;parameter&nbsp;list.&nbsp;Can&nbsp;optionally&nbsp;perform&nbsp;a&nbsp;deep&nbsp;copy<br>
+of&nbsp;the&nbsp;list.</tt></dd></dl>
+
+<dl><dt><a name="Polygon-__str__"><strong>__str__</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;string&nbsp;representation&nbsp;of&nbsp;this&nbsp;<a href="#Polygon">Polygon</a>.&nbsp;Include&nbsp;the&nbsp;first&nbsp;Location<br>
+repeated&nbsp;as&nbsp;the&nbsp;last&nbsp;Location&nbsp;to&nbsp;simulate&nbsp;closedness.</tt></dd></dl>
+
+<dl><dt><a name="Polygon-closed_loop"><strong>closed_loop</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;generator&nbsp;for&nbsp;the&nbsp;Locations&nbsp;in&nbsp;this&nbsp;<a href="#Polygon">Polygon</a>.&nbsp;Will&nbsp;generate&nbsp;the<br>
+first&nbsp;item&nbsp;again&nbsp;at&nbsp;the&nbsp;end,&nbsp;simulating&nbsp;the&nbsp;closedness&nbsp;of&nbsp;the&nbsp;<a href="#Polygon">Polygon</a>.</tt></dd></dl>
+
+<hr>
+Methods inherited from <a href="polyline.html#PolyLine">polyline.PolyLine</a>:<br>
+<dl><dt><a name="Polygon-__eq__"><strong>__eq__</strong></a>(self, other)</dt><dd><tt>Check&nbsp;if&nbsp;this&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>&nbsp;is&nbsp;the&nbsp;same&nbsp;as&nbsp;another&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>.&nbsp;Compares&nbsp;their<br>
+internal&nbsp;Location&nbsp;list.</tt></dd></dl>
+
+<dl><dt><a name="Polygon-compress"><strong>compress</strong></a>(self)</dt><dd><tt>Transform&nbsp;this&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>&nbsp;into&nbsp;its&nbsp;compressed&nbsp;representation.&nbsp;The<br>
+compression&nbsp;is&nbsp;based&nbsp;on&nbsp;the&nbsp;MapQuestion&nbsp;compressed&nbsp;lat/lon&nbsp;encoding<br>
+found&nbsp;here:<br>
+https://developer.mapquest.com/documentation/common/encode-decode/</tt></dd></dl>
+
+<dl><dt><a name="Polygon-get_bounds"><strong>get_bounds</strong></a>(self)</dt><dd><tt>Get&nbsp;the&nbsp;bounding&nbsp;Rectangle&nbsp;of&nbsp;this&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>.</tt></dd></dl>
+
+<dl><dt><a name="Polygon-get_locations_list"><strong>get_locations_list</strong></a>(self)</dt><dd><tt>Get&nbsp;the&nbsp;underlying&nbsp;Location&nbsp;list&nbsp;for&nbsp;this&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>.</tt></dd></dl>
+
+<dl><dt><a name="Polygon-locations"><strong>locations</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;generator&nbsp;for&nbsp;the&nbsp;Locations&nbsp;in&nbsp;this&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>.</tt></dd></dl>
+
+<hr>
+Data descriptors inherited from boundable.Boundable:<br>
+<dl><dt><strong>__dict__</strong></dt>
+<dd><tt>dictionary&nbsp;for&nbsp;instance&nbsp;variables&nbsp;(if&nbsp;defined)</tt></dd>
+</dl>
+<dl><dt><strong>__weakref__</strong></dt>
+<dd><tt>list&nbsp;of&nbsp;weak&nbsp;references&nbsp;to&nbsp;the&nbsp;object&nbsp;(if&nbsp;defined)</tt></dd>
+</dl>
+</td></tr></table></td></tr></table><p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#eeaa77">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Functions</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#eeaa77"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><dl><dt><a name="-decompress_polygon"><strong>decompress_polygon</strong></a>(bytestring)</dt><dd><tt>Given&nbsp;a&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>&nbsp;bytestring&nbsp;compressed&nbsp;using&nbsp;<a href="polyline.html#PolyLine">PolyLine</a>.compress(),<br>
+decompress&nbsp;it&nbsp;and&nbsp;return&nbsp;it&nbsp;as&nbsp;a&nbsp;<a href="#Polygon">Polygon</a>.</tt></dd></dl>
+</td></tr></table>
+</body></html>
