+++ pyatlas/doc/rectangle.html
@@ -10,7 +10,7 @@
 <font color="#ffffff" face="helvetica, arial">&nbsp;<br><big><big><strong>rectangle</strong></big></big></font></td
 ><td align=right valign=bottom
 ><font color="#ffffff" face="helvetica, arial"><a href=".">index</a><br><a href="file:/Users/lucascram/Projects/atlas/pyatlas/pyatlas/rectangle.py">/Users/lucascram/Projects/atlas/pyatlas/pyatlas/rectangle.py</a></font></td></tr></table>
-    <p><tt>Module&nbsp;for&nbsp;a&nbsp;special&nbsp;type&nbsp;of&nbsp;<a href="polyline.html#Polygon">Polygon</a>:&nbsp;the&nbsp;<a href="#Rectangle">Rectangle</a>.</tt></p>
+    <p><tt>Module&nbsp;for&nbsp;a&nbsp;special&nbsp;type&nbsp;of&nbsp;<a href="polygon.html#Polygon">Polygon</a>:&nbsp;the&nbsp;<a href="#Rectangle">Rectangle</a>.</tt></p>
 <p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#aa55cc">
@@ -19,7 +19,7 @@
     
 <tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
 <td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="location.html">location</a><br>
-</td><td width="25%" valign=top><a href="polyline.html">polyline</a><br>
+</td><td width="25%" valign=top><a href="polygon.html">polygon</a><br>
 </td><td width="25%" valign=top></td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#ee77aa">
@@ -28,7 +28,7 @@
     
 <tr><td bgcolor="#ee77aa"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
 <td width="100%"><dl>
-<dt><font face="helvetica, arial"><a href="polyline.html#Polygon">polyline.Polygon</a>(<a href="polyline.html#PolyLine">polyline.PolyLine</a>)
+<dt><font face="helvetica, arial"><a href="polygon.html#Polygon">polygon.Polygon</a>(polyline.PolyLine)
 </font></dt><dd>
 <dl>
 <dt><font face="helvetica, arial"><a href="rectangle.html#Rectangle">Rectangle</a>
@@ -39,7 +39,7 @@
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#ffc8d8">
 <td colspan=3 valign=bottom>&nbsp;<br>
-<font color="#000000" face="helvetica, arial"><a name="Rectangle">class <strong>Rectangle</strong></a>(<a href="polyline.html#Polygon">polyline.Polygon</a>)</font></td></tr>
+<font color="#000000" face="helvetica, arial"><a name="Rectangle">class <strong>Rectangle</strong></a>(<a href="polygon.html#Polygon">polygon.Polygon</a>)</font></td></tr>
     
 <tr bgcolor="#ffc8d8"><td rowspan=2><tt>&nbsp;&nbsp;&nbsp;</tt></td>
 <td colspan=2><tt>A&nbsp;rectangle&nbsp;on&nbsp;the&nbsp;surface&nbsp;of&nbsp;earth.&nbsp;It&nbsp;cannot&nbsp;span&nbsp;the&nbsp;date&nbsp;change<br>
@@ -47,9 +47,9 @@ line&nbsp;(longitude&nbsp;-180).<br>&nbsp;</tt></td></tr>
 <tr><td>&nbsp;</td>
 <td width="100%"><dl><dt>Method resolution order:</dt>
 <dd><a href="rectangle.html#Rectangle">Rectangle</a></dd>
-<dd><a href="polyline.html#Polygon">polyline.Polygon</a></dd>
-<dd><a href="polyline.html#PolyLine">polyline.PolyLine</a></dd>
-<dd><a href="boundable.html#Boundable">boundable.Boundable</a></dd>
+<dd><a href="polygon.html#Polygon">polygon.Polygon</a></dd>
+<dd>polyline.PolyLine</dd>
+<dd>boundable.Boundable</dd>
 <dd><a href="__builtin__.html#object">__builtin__.object</a></dd>
 </dl>
 <hr>
@@ -62,15 +62,15 @@ upper&nbsp;right&nbsp;corner&nbsp;Location.</tt></dd></dl>
 <dl><dt><a name="Rectangle-get_upper_right"><strong>get_upper_right</strong></a>(self)</dt><dd><tt>Get&nbsp;the&nbsp;upper&nbsp;right&nbsp;corner&nbsp;Location&nbsp;of&nbsp;this&nbsp;<a href="#Rectangle">Rectangle</a>.</tt></dd></dl>
 
 <hr>
-Methods inherited from <a href="polyline.html#Polygon">polyline.Polygon</a>:<br>
-<dl><dt><a name="Rectangle-__str__"><strong>__str__</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;string&nbsp;representation&nbsp;of&nbsp;this&nbsp;<a href="polyline.html#Polygon">Polygon</a>.&nbsp;Include&nbsp;the&nbsp;first&nbsp;Location<br>
+Methods inherited from <a href="polygon.html#Polygon">polygon.Polygon</a>:<br>
+<dl><dt><a name="Rectangle-__str__"><strong>__str__</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;string&nbsp;representation&nbsp;of&nbsp;this&nbsp;<a href="polygon.html#Polygon">Polygon</a>.&nbsp;Include&nbsp;the&nbsp;first&nbsp;Location<br>
 repeated&nbsp;as&nbsp;the&nbsp;last&nbsp;Location&nbsp;to&nbsp;simulate&nbsp;closedness.</tt></dd></dl>
 
-<dl><dt><a name="Rectangle-closed_loop"><strong>closed_loop</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;generator&nbsp;for&nbsp;the&nbsp;Locations&nbsp;in&nbsp;this&nbsp;<a href="polyline.html#Polygon">Polygon</a>.&nbsp;Will&nbsp;generate&nbsp;the<br>
-first&nbsp;item&nbsp;again&nbsp;at&nbsp;the&nbsp;end,&nbsp;simulating&nbsp;the&nbsp;closedness&nbsp;of&nbsp;the&nbsp;<a href="polyline.html#Polygon">Polygon</a>.</tt></dd></dl>
+<dl><dt><a name="Rectangle-closed_loop"><strong>closed_loop</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;generator&nbsp;for&nbsp;the&nbsp;Locations&nbsp;in&nbsp;this&nbsp;<a href="polygon.html#Polygon">Polygon</a>.&nbsp;Will&nbsp;generate&nbsp;the<br>
+first&nbsp;item&nbsp;again&nbsp;at&nbsp;the&nbsp;end,&nbsp;simulating&nbsp;the&nbsp;closedness&nbsp;of&nbsp;the&nbsp;<a href="polygon.html#Polygon">Polygon</a>.</tt></dd></dl>
 
 <hr>
-Methods inherited from <a href="polyline.html#PolyLine">polyline.PolyLine</a>:<br>
+Methods inherited from polyline.PolyLine:<br>
 <dl><dt><a name="Rectangle-__eq__"><strong>__eq__</strong></a>(self, other)</dt><dd><tt>Check&nbsp;if&nbsp;this&nbsp;PolyLine&nbsp;is&nbsp;the&nbsp;same&nbsp;as&nbsp;another&nbsp;PolyLine.&nbsp;Compares&nbsp;their<br>
 internal&nbsp;Location&nbsp;list.</tt></dd></dl>
 
@@ -86,7 +86,7 @@ https://developer.mapquest.com/documentation/common/encode-decode/</tt></dd></dl
 <dl><dt><a name="Rectangle-locations"><strong>locations</strong></a>(self)</dt><dd><tt>Get&nbsp;a&nbsp;generator&nbsp;for&nbsp;the&nbsp;Locations&nbsp;in&nbsp;this&nbsp;PolyLine.</tt></dd></dl>
 
 <hr>
-Data descriptors inherited from <a href="boundable.html#Boundable">boundable.Boundable</a>:<br>
+Data descriptors inherited from boundable.Boundable:<br>
 <dl><dt><strong>__dict__</strong></dt>
 <dd><tt>dictionary&nbsp;for&nbsp;instance&nbsp;variables&nbsp;(if&nbsp;defined)</tt></dd>
 </dl>
